+++ Sources/BraintreePayPal/BTPayPalClient.swift
@@ -69,11 +69,11 @@ import BraintreeDataCollector
 
     /// Initialize a new PayPal client instance.
     /// - Parameter apiClient: The API Client
-    @objc(initWithAPIClient:)
-    public init(apiClient: BTAPIClient) {
+    @objc(initWithAuthorization:)
+    public init(authorization: String) {
         BTAppContextSwitcher.sharedInstance.register(BTPayPalClient.self)
 
-        self.apiClient = apiClient
+        self.apiClient = BTAPIClient(newAuthorization: authorization)
         self.webAuthenticationSession = BTWebAuthenticationSession()
 
         super.init()
@@ -90,9 +90,9 @@ import BraintreeDataCollector
     ///   - apiClient: The API Client
     ///   - universalLink: The URL to use for the PayPal app switch flow. Must be a valid HTTPS URL dedicated to Braintree app switch returns. This URL must be allow-listed in your Braintree Control Panel.
     /// - Warning: This initializer should be used for merchants using the PayPal App Switch flow. This feature is currently in beta and may change or be removed in future releases.
-    @objc(initWithAPIClient:universalLink:)
-    public convenience init(apiClient: BTAPIClient, universalLink: URL) {
-        self.init(apiClient: apiClient)
+    @objc(initWithAuthorization:universalLink:)
+    public convenience init(authorization: String, universalLink: URL) {
+        self.init(authorization: authorization)
         self.universalLink = universalLink
     }