+++ CHANGELOG.md
@@ -31,6 +31,8 @@
     * Update PayPal app URL query scheme from `paypal-app-switch-checkout` to `paypal`
   * BraintreeAmericanExpress
     * Update initializer to `BTAmericanExpressClient(authorization:)`
+  * BraintreePayPal
+    * Update initializer to `BTPayPalClient(authorization:)`
 
 ## 6.27.0 (2025-01-23)
 * BraintreePayPal